Add Verficiation of hash, update structure
verifiy the hash of the access token and store it in the according map.
change the overall structure to be always the same:
each token is represented as a map with at least one key 'token'
which contains the raw, binary token.

The token maps are values inside a map, each accessable with a key
corresponding to their name. id => IdTokenMap, access => AccessTokenMap
and refresh => RefreshTokenMap.