feat: create an httpc profile which disables keep-alive/pipelining
As noted in
https://github.com/erlef/oidcc/issues/318#issuecomment-1880065555,
`httpc` by default will re-use existing connections. This is great if
you're using normal HTTPS, but if you're using client authentication
then you need to make sure that every time `httpc` connects to a host,
it's using the client authentication, which is impossible in practice.

This works around that, by creating a new profile which disables that
functionality. Using that profile for requests which provide SSL
overrides will ensure that each of those requests will use the client certificate.